+++ Example/Example2_RandomizedSectionsAnimation.swift
@@ -35,7 +35,7 @@ class ViewController: UIViewController {
         let initialRandomizedSections = Randomizer(rng: PseudoRandomGenerator(4, 3), sections: initialValue())
 
         let ticks = Observable<Int>.interval(1, scheduler: MainScheduler.instance).map { _ in () }
-        let randomSections = Observable.of(ticks, refreshButton.rx_tap.asObservable())
+        let randomSections = Observable.of(ticks, refreshButton.rx.tap.asObservable())
                 .merge()
                 .scan(initialRandomizedSections) { a, _ in
                     return a.randomize()
@@ -51,51 +51,31 @@ class ViewController: UIViewController {
         skinTableViewDataSource(reloadDataSource)
 
         randomSections
-            .bindTo(animatedTableView.rx_itemsWithDataSource(tvAnimatedDataSource))
+            .bindTo(animatedTableView.rx.items(dataSource: tvAnimatedDataSource))
             .addDisposableTo(disposeBag)
 
         randomSections
-            .bindTo(tableView.rx_itemsWithDataSource(reloadDataSource))
+            .bindTo(tableView.rx.items(dataSource: reloadDataSource))
             .addDisposableTo(disposeBag)
 
-        // Collection view logic works, but when clicking fast because of internal bugs
-        // collection view will sometimes get confused. I know what you are thinking,
-        // but this is really not a bug in the algorithm. The generated changes are
-        // pseudorandom, and crash happens depending on clicking speed.
-        //
-        // If you want, turn this to true, just click slow :)
-        //
-        // While `useAnimatedUpdateForCollectionView` is false, you can click as fast as
-        // you want, table view doesn't seem to have same issues like collection view.
+        let cvAnimatedDataSource = RxCollectionViewSectionedAnimatedDataSource<NumberSection>()
+        skinCollectionViewDataSource(cvAnimatedDataSource)
 
-        let useAnimatedUpdates = true
-        if useAnimatedUpdates {
-            let cvAnimatedDataSource = RxCollectionViewSectionedAnimatedDataSource<NumberSection>()
-            skinCollectionViewDataSource(cvAnimatedDataSource)
-
-            randomSections
-                .bindTo(animatedCollectionView.rx_itemsWithDataSource(cvAnimatedDataSource))
-                .addDisposableTo(disposeBag)
-        }
-        else {
-            let cvReloadDataSource = RxCollectionViewSectionedReloadDataSource<NumberSection>()
-            skinCollectionViewDataSource(cvReloadDataSource)
-            randomSections
-                .bindTo(animatedCollectionView.rx_itemsWithDataSource(cvReloadDataSource))
-                .addDisposableTo(disposeBag)
-        }
+        randomSections
+            .bindTo(animatedCollectionView.rx.items(dataSource: cvAnimatedDataSource))
+            .addDisposableTo(disposeBag)
 
         // touches
 
         Observable.of(
-            tableView.rx_modelSelected(IntItem.self),
-            animatedTableView.rx_modelSelected(IntItem.self),
-            animatedCollectionView.rx_modelSelected(IntItem.self)
+            tableView.rx.modelSelected(IntItem.self),
+            animatedTableView.rx.modelSelected(IntItem.self),
+            animatedCollectionView.rx.modelSelected(IntItem.self)
         )
             .merge()
-            .subscribeNext { item in
+            .subscribe(onNext: { item in
                 print("Let me guess, it's .... It's \(item), isn't it? Yeah, I've got it.")
-            }
+            })
             .addDisposableTo(disposeBag)
     }
 }
@@ -103,9 +83,9 @@ class ViewController: UIViewController {
 // MARK: Skinning
 extension ViewController {
 
-    func skinTableViewDataSource(dataSource: RxTableViewSectionedDataSource<NumberSection>) {
+    func skinTableViewDataSource(_ dataSource: TableViewSectionedDataSource<NumberSection>) {
         dataSource.configureCell = { (_, tv, ip, i) in
-            let cell = tv.dequeueReusableCellWithIdentifier("Cell") ?? UITableViewCell(style:.Default, reuseIdentifier: "Cell")
+            let cell = tv.dequeueReusableCell(withIdentifier: "Cell") ?? UITableViewCell(style:.default, reuseIdentifier: "Cell")
 
             cell.textLabel!.text = "\(i)"
 
@@ -113,13 +93,13 @@ extension ViewController {
         }
 
         dataSource.titleForHeaderInSection = { (ds, section) -> String? in
-            return ds.sectionAtIndex(section).header
+            return ds[section].header
         }
     }
 
-    func skinCollectionViewDataSource(dataSource: CollectionViewSectionedDataSource<NumberSection>) {
+    func skinCollectionViewDataSource(_ dataSource: CollectionViewSectionedDataSource<NumberSection>) {
         dataSource.configureCell = { (_, cv, ip, i) in
-            let cell = cv.dequeueReusableCellWithReuseIdentifier("Cell", forIndexPath: ip) as! NumberCell
+            let cell = cv.dequeueReusableCell(withReuseIdentifier: "Cell", for: ip) as! NumberCell
 
             cell.value!.text = "\(i)"
 
@@ -127,9 +107,9 @@ extension ViewController {
         }
 
         dataSource.supplementaryViewFactory = { (ds ,cv, kind, ip) in
-            let section = cv.dequeueReusableSupplementaryViewOfKind(kind, withReuseIdentifier: "Section", forIndexPath: ip) as! NumberSectionView
+            let section = cv.dequeueReusableSupplementaryView(ofKind: kind, withReuseIdentifier: "Section", for: ip) as! NumberSectionView
 
-            section.value!.text = "\(ds.sectionAtIndex(ip.section).header)"
+            section.value!.text = "\(ds[ip.section].header)"
             
             return section
         }
@@ -138,11 +118,9 @@ extension ViewController {
     // MARK: Initial value
 
     func initialValue() -> [NumberSection] {
-        let generate = true
-        if generate {
-
-            let nSections = 4
-            let nItems = 2
+        #if true
+            let nSections = 10
+            let nItems = 100
 
 
             /*
@@ -151,31 +129,30 @@ extension ViewController {
             */
 
             return (0 ..< nSections).map { (i: Int) in
-                NumberSection(header: "Section \(i + 1)", numbers: $(Array(i * nItems ..< (i + 1) * nItems)), updated: NSDate())
+                NumberSection(header: "Section \(i + 1)", numbers: $(Array(i * nItems ..< (i + 1) * nItems)), updated: Date())
             }
-        }
-        else {
+        #else
             return _initialValue
-        }
+        #endif
     }
 
 
 }
 
 let _initialValue: [NumberSection] = [
-    NumberSection(header: "section 1", numbers: $([1, 2, 3]), updated: NSDate()),
-    NumberSection(header: "section 2", numbers: $([4, 5, 6]), updated: NSDate()),
-    NumberSection(header: "section 3", numbers: $([7, 8, 9]), updated: NSDate()),
-    NumberSection(header: "section 4", numbers: $([10, 11, 12]), updated: NSDate()),
-    NumberSection(header: "section 5", numbers: $([13, 14, 15]), updated: NSDate()),
-    NumberSection(header: "section 6", numbers: $([16, 17, 18]), updated: NSDate()),
-    NumberSection(header: "section 7", numbers: $([19, 20, 21]), updated: NSDate()),
-    NumberSection(header: "section 8", numbers: $([22, 23, 24]), updated: NSDate()),
-    NumberSection(header: "section 9", numbers: $([25, 26, 27]), updated: NSDate()),
-    NumberSection(header: "section 10", numbers: $([28, 29, 30]), updated: NSDate())
+    NumberSection(header: "section 1", numbers: $([1, 2, 3]), updated: Date()),
+    NumberSection(header: "section 2", numbers: $([4, 5, 6]), updated: Date()),
+    NumberSection(header: "section 3", numbers: $([7, 8, 9]), updated: Date()),
+    NumberSection(header: "section 4", numbers: $([10, 11, 12]), updated: Date()),
+    NumberSection(header: "section 5", numbers: $([13, 14, 15]), updated: Date()),
+    NumberSection(header: "section 6", numbers: $([16, 17, 18]), updated: Date()),
+    NumberSection(header: "section 7", numbers: $([19, 20, 21]), updated: Date()),
+    NumberSection(header: "section 8", numbers: $([22, 23, 24]), updated: Date()),
+    NumberSection(header: "section 9", numbers: $([25, 26, 27]), updated: Date()),
+    NumberSection(header: "section 10", numbers: $([28, 29, 30]), updated: Date())
 ]
 
-func $(numbers: [Int]) -> [IntItem] {
-    return numbers.map { IntItem(number: $0, date: NSDate()) }
+func $(_ numbers: [Int]) -> [IntItem] {
+    return numbers.map { IntItem(number: $0, date: Date()) }
 }