+++ Example/Example4_DifferentSectionAndItemTypes.swift
@@ -34,13 +34,13 @@ class MultipleSectionModelViewController: UIViewController {
         skinTableViewDataSource(dataSource)
         
         Observable.just(sections)
-            .bindTo(tableView.rx_itemsWithDataSource(dataSource))
+            .bindTo(tableView.rx.items(dataSource: dataSource))
             .addDisposableTo(disposeBag)
     }
     
-    func skinTableViewDataSource(dataSource: RxTableViewSectionedReloadDataSource<MultipleSectionModel>) {
+    func skinTableViewDataSource(_ dataSource: RxTableViewSectionedReloadDataSource<MultipleSectionModel>) {
         dataSource.configureCell = { (dataSource, table, idxPath, _) in
-            switch dataSource.itemAtIndexPath(idxPath) {
+            switch dataSource[idxPath] {
             case let .ImageSectionItem(image, title):
                 let cell: ImageTitleTableViewCell = table.dequeueReusableCell(forIndexPath: idxPath)
                 cell.titleLabel.text = title
@@ -54,7 +54,7 @@ class MultipleSectionModelViewController: UIViewController {
                 return cell
             case let .ToggleableSectionItem(title, enabled):
                 let cell: TitleSwitchTableViewCell = table.dequeueReusableCell(forIndexPath: idxPath)
-                cell.switchControl.on = enabled
+                cell.switchControl.isOn = enabled
                 cell.titleLabel.text = title
                 
                 return cell
@@ -62,7 +62,7 @@ class MultipleSectionModelViewController: UIViewController {
         }
 
         dataSource.titleForHeaderInSection = { dataSource, index in
-            let section = dataSource.sectionAtIndex(index)
+            let section = dataSource[index]
             
             return section.title
         }