+++ Example/Views/UIKitExtensions.swift
@@ -8,7 +8,7 @@
 
 import class UIKit.UITableViewCell
 import class UIKit.UITableView
-import class Foundation.NSIndexPath
+import struct Foundation.IndexPath
 
 protocol ReusableView: class {
     static var reuseIdentifier: String {get}
@@ -16,7 +16,7 @@ protocol ReusableView: class {
 
 extension ReusableView {
     static var reuseIdentifier: String {
-        return String(self)
+        return String(describing: self)
     }
 }
 
@@ -25,8 +25,8 @@ extension UITableViewCell: ReusableView {
 
 extension UITableView {
     
-    func dequeueReusableCell<T: UITableViewCell where T: ReusableView>(forIndexPath indexPath: NSIndexPath) -> T {
-        guard let cell = dequeueReusableCellWithIdentifier(T.reuseIdentifier, forIndexPath: indexPath) as? T else {
+    func dequeueReusableCell<T: UITableViewCell>(forIndexPath indexPath: IndexPath) -> T where T: ReusableView {
+        guard let cell = dequeueReusableCell(withIdentifier: T.reuseIdentifier, for: indexPath) as? T else {
             fatalError("Could not dequeue cell with identifier: \(T.reuseIdentifier)")
         }